s/tools/ctlcmd/
