+++ overlord/hookstate/ctlcmd/ctlcmd_test.go
@@ -17,52 +17,52 @@
  *
  */
 
-package tools_test
+package ctlcmd_test
 
 import (
 	"testing"
 
+	"github.com/snapcore/snapd/overlord/hookstate/ctlcmd"
 	"github.com/snapcore/snapd/overlord/hookstate/hooktest"
-	"github.com/snapcore/snapd/overlord/hookstate/tools"
 
 	. "gopkg.in/check.v1"
 )
 
 func Test(t *testing.T) { TestingT(t) }
 
-type toolsSuite struct {
+type ctlcmdSuite struct {
 	mockHandler *hooktest.MockHandler
 }
 
-var _ = Suite(&toolsSuite{})
+var _ = Suite(&ctlcmdSuite{})
 
-func (s *toolsSuite) SetUpTest(c *C) {
+func (s *ctlcmdSuite) SetUpTest(c *C) {
 	s.mockHandler = hooktest.NewMockHandler()
 }
 
-func (s *toolsSuite) TestNonExistingCommand(c *C) {
-	stdout, stderr, err := tools.RunCommand(s.mockHandler, []string{"foo", "--bar"})
+func (s *ctlcmdSuite) TestNonExistingCommand(c *C) {
+	stdout, stderr, err := ctlcmd.RunCommand(s.mockHandler, []string{"foo", "--bar"})
 	c.Check(string(stdout), Equals, "")
 	c.Check(string(stderr), Equals, "")
 	c.Check(err, ErrorMatches, ".*[Uu]nknown command.*")
 }
 
-func (s *toolsSuite) TestCommandOutput(c *C) {
-	mockCommand := tools.NewMockCommand()
+func (s *ctlcmdSuite) TestCommandOutput(c *C) {
+	mockCommand := ctlcmd.NewMockCommand()
 	mockCommand.FakeStdout = "test stdout"
 	mockCommand.FakeStderr = "test stderr"
-	tools.AddCommand("mock", mockCommand)
-	defer tools.RemoveCommand("mock")
+	ctlcmd.AddCommand("mock", mockCommand)
+	defer ctlcmd.RemoveCommand("mock")
 
-	stdout, stderr, err := tools.RunCommand(s.mockHandler, []string{"mock", "foo", "--bar"})
+	stdout, stderr, err := ctlcmd.RunCommand(s.mockHandler, []string{"mock", "foo", "--bar"})
 	c.Check(err, IsNil)
 	c.Check(string(stdout), Equals, "test stdout")
 	c.Check(string(stderr), Equals, "test stderr")
 	c.Check(mockCommand.Args, DeepEquals, []string{"foo", "--bar"})
 }
 
-func (s *toolsSuite) TestSetCommand(c *C) {
-	stdout, stderr, err := tools.RunCommand(s.mockHandler, []string{"set", "foo=bar"})
+func (s *ctlcmdSuite) TestSetCommand(c *C) {
+	stdout, stderr, err := ctlcmd.RunCommand(s.mockHandler, []string{"set", "foo=bar"})
 	c.Check(err, IsNil)
 	c.Check(string(stdout), Equals, "")
 	c.Check(string(stderr), Equals, "")
